Multimedia key fix
"There's a bug with the fix I added to the double multimedia key input bug. It sometimes doesnt work.
The alt key would randomly be pressed, even though it's not, and that would mess up the multimedia key.
Now it doesnt, but pressing multiple multimedia key at the same time causes them to not work afterwards, until a non multimedia key gets pressed."